requirements.txt
asgiref==3.8.1
awscli==1.35.0
awsebcli==3.21.0
blessed==1.20.0
botocore==1.35.34
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
docutils==0.16
idna==3.10
jmespath==1.0.1
pathspec==0.10.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
semantic-version==2.10.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
termcolor==2.5.0
urllib3==1.26.20
wcwidth==0.2.13
